Move memory allocating utilities to 'regions'
This adds a new subdirectory 'regions' to the library that contains the
miscellaneous memory regions used for allocations in this library
