+++ include/bit/memory/allocators/aligned_offset_allocator.hpp
@@ -11,10 +11,11 @@
 
 #include "detail/named_allocator.hpp" // detail::named_allocator
 
-#include "../aligned_heap_memory.hpp" // aligned_malloc, aligned_free
 #include "../allocator_info.hpp" // allocator_info
 #include "../macros.hpp"         // BIT_MEMORY_UNUSED
 #include "../owner.hpp"          // owner
+
+#include "../regions/aligned_heap_memory.hpp" // aligned_malloc, aligned_free
 
 #include <cstddef>     // std::max_align_t
 #include <type_traits> // std::true_type
