+++ src/bit/memory/regions/posix/aligned_heap_memory.cpp
@@ -1,4 +1,4 @@
-#include <bit/memory/aligned_heap_memory.hpp>
+#include <bit/memory/regions/aligned_heap_memory.hpp>
 #include <bit/memory/pointer_utilities.hpp> // offset_align_forward
 #include <bit/memory/unaligned_storage.hpp>  // store_unaligned, load_unaligned
